# edge-control-plane/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: edgeCloud Control Plane API
  version: 1.0.0
  description: |
    HTTP API for the edgeCloud managed WebAssembly edge computing platform.
    Supports tenant and worker authentication. Tenants authenticate via
    Bearer API key. Workers authenticate via Bearer JWT (HMAC-SHA256).
servers:
  - url: https://api.edgecloud.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Tenant API key (Bearer token). Obtain via POST /api/v1/keys.
    WorkerJWTAuth:
      type: http
      scheme: bearer
      description: Worker JWT (HMAC-SHA256). Issued to workers by the control plane.
  schemas:
...
    BootstrapRequest:
      type: object
      required: [name, key_name]
      properties:
        name:
          type: string
          description: Unique tenant name.
          example: acme
        plan:
          type: string
          default: free
          description: Subscription plan.
          example: free
        key_name:
          type: string
          description: Name for the first API key.
          example: default
    BootstrapResponse:
      type: object
      properties:
        tenant_id:
          type: string
          description: The new tenant's ID.
          example: t_abc123
        api_key:
          type: string
          description: |
            The raw API key. Shown only once — store it securely.
            Use Authorization: Bearer <api_key> for all subsequent requests.
          example: ek_...
...
    ActivateResponse:
      type: object
      properties:
        status:
          type: string
          example: activated
...
paths:
...
  /api/v1/tenants:
    post:
      operationId: bootstrapTenant
      summary: Self-signup — create tenant and first API key
      description: |
        Creates a new tenant and atomically issues the first API key.
        No authentication required. The returned raw API key is shown only once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
            example:
              name: acme
              key_name: default
      responses:
        '201':
          description: Tenant and API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/keys:
    post:
      operationId: createAPIKey
      summary: Create a new API key for the authenticated tenant
      description: |
        Issues a new API key scoped to the authenticated tenant.
        The raw key is returned only once — store it securely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
            example:
              name: ci-key
              role: developer
      responses:
        '201':
          description: API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listAPIKeys
      summary: List all API keys for the authenticated tenant
      responses:
        '200':
          description: List of API key metadata (raw keys not included).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Auth
  # -------------------------------------------------------------------------
  /api/v1/auth/whoami:
    get:
      operationId: whoami
      summary: Return the authenticated tenant and API key identity
      responses:
        '200':
          description: Authenticated identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhoamiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Apps
  # -------------------------------------------------------------------------
  /api/v1/apps:
    get:
      operationId: listApps
      summary: List all apps for the authenticated tenant
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of apps.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/apps/{appName}:
    post:
      operationId: createApp
      summary: Create a new app
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppRequest'
      responses:
        '201':
          description: App created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/QuotaExceeded'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: getApp
      summary: Get a specific app
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
      responses:
        '200':
          description: App details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Deployments
  # -------------------------------------------------------------------------
  /api/v1/deploy/{appName}:
    post:
      operationId: deploy
      summary: Upload and deploy a Wasm artifact
      description: |
        Stores the artifact, registers a new deployment, and auto-creates
        the app record if it does not already exist. The artifact is
        validated as a WebAssembly binary before storage.
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: Raw Wasm artifact bytes.
      responses:
        '201':
          description: Deployment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/QuotaExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/status/{deploymentID}:
    get:
      operationId: getDeploymentStatus
      summary: Get deployment details by ID
      parameters:
        - name: deploymentID
          in: path
          required: true
          schema:
            type: string
          description: The deployment ID returned by the deploy endpoint.
      responses:
        '200':
          description: Deployment details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/list/{appName}:
    get:
      operationId: listDeployments
      summary: List deployments for an app with pagination
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated deployment list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/apps/{appName}/activate/{deploymentID}:
    post:
      operationId: activateDeployment
      summary: Activate a specific deployment as the live version
      description: |
        Publishes a task update over NATS JetStream so subscribed workers
        start serving the specified deployment. Only one deployment per app
        can be active at a time — activating a new one replaces the current
        live deployment atomically.
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
        - name: deploymentID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/apps/{appName}/rollback:
    post:
      operationId: rollbackDeployment
      summary: Roll back the active deployment to the stored last-good deployment
      description: |
        Replaces the currently active deployment of `appName` with the
        deployment recorded as the most recent `last_good_deployment_id`.
        Used after a bad activation to put the app back on a known-good
        version. The previously active deployment becomes the new
        `last_good_deployment_id`, so the operation is itself reversible.
        A NATS task update is published to every region the deployment
        is replicated to so workers re-converge.
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
      responses:
        '200':
          description: Deployment rolled back.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RollbackResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            No prior `last_good_deployment_id` is recorded for this app,
            so there is nothing to roll back to. Returned when the app
            has only ever had one deployment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: |
            The post-commit NATS publish of the task update failed for
            every region the deployment is replicated to. The control
            plane has already committed the new active deployment, but
            workers may not converge until the publish is retried.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/apps/{appName}/active:
    get:
      operationId: getActiveDeployment
      summary: Get the currently active deployment for an app
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
      responses:
        '200':
          description: Active deployment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/apps/{appName}/logs:
    get:
      operationId: listLogsForApp
      summary: List recent log entries for an app
      description: |
        Returns the most recent log entries for the named app within
        the caller's tenant, newest first. Supports `since` (RFC3339
        cutoff), `level` (minimum severity), and `limit` (max items).

        The cutoff is computed server-side against the database clock
        (NOW() - interval), so wall-clock skew between this server
        and the database does not affect what is returned.
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            RFC3339 cutoff; missing → server default (5m). Future-dated
            values are clamped to the default window.
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [trace, debug, info, warn, error]
          description: |
            Minimum severity. Filter is inclusive: `level=warn` returns
            `warn` and `error`. Unknown values → 400.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: |
            Maximum items to return. Values outside [1, 1000] are
            clamped; absent or zero → server default (100).
      responses:
        '200':
          description: Log entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/apps/{appName}/env/{key}:
    delete:
      operationId: deleteEnvVar
      summary: Delete an environment variable
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          description: Unique name of the app within the tenant.
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: The environment variable key to delete.
      responses:
        '204':
          description: Environment variable deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Migration
  # -------------------------------------------------------------------------
...
  /api/v1/quotas:
    get:
      operationId: getQuota
      summary: Get the authenticated tenant's resource quota limits
      responses:
        '200':
          description: Quota limits and current usage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quota'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...